还在尝试实现查询功能,使用中间表实体类策略，正在进行中

+++ src/main/java/cn/wowtw_backend/repository/InfoGroupRepository.java
@@ -1,14 +1,16 @@
 package cn.wowtw_backend.repository;
 
+import cn.wowtw_backend.model.infoGroup.GearExtra;
 import cn.wowtw_backend.model.infoGroup.InfoGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface InfoGroupRepository extends JpaRepository<InfoGroup, Byte> {
-//    @Query("select t FROM TalentInfoGroup t WHERE t.id = :talentId")
-//    List<InfoGroup> findInfoGroupByTalentId(Integer talentId);
+public interface InfoGroupRepository extends JpaRepository<InfoGroup, Integer> {
+
     @Query("SELECT ig FROM InfoGroup ig JOIN TalentInfoGroup tig ON ig.id = tig.infoGroup.id WHERE tig.talent.id = :talentId")
     List<InfoGroup> findInfoGroupByTalentId(Integer talentId);
+
+
 }
